+++ tests/PointInTimeTest.php
@@ -36,7 +36,7 @@ class PointInTimeTest extends TestCase
         $this->assertInstanceOf(Offset::class, $point->offset());
         $this->assertSame(2016, $point->year()->toInt());
         $this->assertSame(10, $point->month()->ofYear()->toInt());
-        $this->assertSame(5, $point->day()->toInt());
+        $this->assertSame(5, $point->day()->ofMonth());
         $this->assertSame(8, $point->hour()->toInt());
         $this->assertSame(1, $point->minute()->toInt());
         $this->assertSame(30, $point->second()->toInt());
@@ -83,7 +83,7 @@ class PointInTimeTest extends TestCase
         $this->assertSame($point->milliseconds(), $point2->milliseconds());
         $this->assertSame(2016, $point2->year()->toInt());
         $this->assertSame(10, $point2->month()->ofYear()->toInt());
-        $this->assertSame(5, $point2->day()->toInt());
+        $this->assertSame(5, $point2->day()->ofMonth());
         $this->assertSame(3, $point2->hour()->toInt());
         $this->assertSame(31, $point2->minute()->toInt());
         $this->assertSame(30, $point2->second()->toInt());
@@ -129,14 +129,14 @@ class PointInTimeTest extends TestCase
 
         $this->assertSame(2016, $point->year()->toInt());
         $this->assertSame(10, $point->month()->ofYear()->toInt());
-        $this->assertSame(5, $point->day()->toInt());
+        $this->assertSame(5, $point->day()->ofMonth());
         $this->assertSame(8, $point->hour()->toInt());
         $this->assertSame(1, $point->minute()->toInt());
         $this->assertSame(30, $point->second()->toInt());
         $this->assertSame(123, $point->millisecond()->toInt());
         $this->assertSame(2017, $point2->year()->toInt());
         $this->assertSame(11, $point2->month()->ofYear()->toInt());
-        $this->assertSame(6, $point2->day()->toInt());
+        $this->assertSame(6, $point2->day()->ofMonth());
         $this->assertSame(9, $point2->hour()->toInt());
         $this->assertSame(3, $point2->minute()->toInt());
         $this->assertSame(1, $point2->second()->toInt());
@@ -152,14 +152,14 @@ class PointInTimeTest extends TestCase
 
         $this->assertSame(2016, $point->year()->toInt());
         $this->assertSame(10, $point->month()->ofYear()->toInt());
-        $this->assertSame(5, $point->day()->toInt());
+        $this->assertSame(5, $point->day()->ofMonth());
         $this->assertSame(8, $point->hour()->toInt());
         $this->assertSame(1, $point->minute()->toInt());
         $this->assertSame(30, $point->second()->toInt());
         $this->assertSame(123, $point->millisecond()->toInt());
         $this->assertSame(2015, $point2->year()->toInt());
         $this->assertSame(9, $point2->month()->ofYear()->toInt());
-        $this->assertSame(4, $point2->day()->toInt());
+        $this->assertSame(4, $point2->day()->ofMonth());
         $this->assertSame(6, $point2->hour()->toInt());
         $this->assertSame(59, $point2->minute()->toInt());
         $this->assertSame(59, $point2->second()->toInt());